requirements-airflow.txt
apache-airflow >= 3.0.2
numpy>=2.3.0
pandas>=2.3.0
polars>=1.30.0
scikit-learn>=1.7.0
dvc >= 3.60.1
